Crate dfn_http

Modules

types

Structs

Header

Request

RequestWrapper
This is currently unstable and may stop working on the IC with little or no warning. This represents a HTTP request and accompanying metadata

Response

Enums

Method

Constants

HTTP_CANISTER_ID

Functions

http
This is currently unstable and may stop working on the IC with little or no warning

http_unchecked
This is largely the same as "http" but the caller ID isn't checked This allows other canisters to call this endpoint with arbitrary data. This could lead to security issues if you're using connection metadata to identify users. If you're not sure what function to use, use "http"