Function dfn_http::http

pub fn http<F>(handler: F) where
    F: FnOnce(RequestWrapper) -> Response, 

This is currently unstable and may stop working on the IC with little or no warning

This allows you to use a canister as an HTTP server

Example

#[export_name = "canister_query http_query"]
pub fn main() {
    http(handler)
}

fn handler(http: RequestWrapper) -> Response {
// Take the request and return a response
}

Requests send to query.<canister_id>.<ic_domain> will reach the handler